Fix: don't display tags of non existing projects
Non existing projects should not be visible (except from search),
and so tags attached to only these projects should not be shown into the
tag list.

So these tags should be removed from jtags, but we would like to keep them
in case of an error or a revival of the project.